+++ apps/fluent-tester/src/E2E/common/NavigateAppPage.ts
@@ -20,7 +20,9 @@ import { HOMEPAGE_PERSONA_BUTTON } from '../../TestComponents/Persona/consts';
 import { HOMEPAGE_PERSONACOIN_BUTTON } from '../../TestComponents/PersonaCoin/consts';
 import { HOMEPAGE_PRESSABLE_BUTTON } from '../../TestComponents/Pressable/consts';
 import { HOMEPAGE_RADIOGROUP_BUTTON } from '../../TestComponents/RadioGroup/consts';
+import { HOMEPAGE_RADIO_GROUP_EXPERIMENTAL_BUTTON } from '../../TestComponents/RadioGroupExperimental/consts';
 import { HOMEPAGE_SEPARATOR_BUTTON } from '../../TestComponents/Separator/consts';
+import { HOMEPAGE_SHADOW_BUTTON } from '../../TestComponents/Shadow/consts';
 import { HOMEPAGE_SHIMMER_BUTTON } from '../../TestComponents/Shimmer/consts';
 import { HOMEPAGE_STROKEWIDTH_BUTTON } from '../../TestComponents/StrokeWidth/consts';
 import { HOMEPAGE_SVG_BUTTON } from '../../TestComponents/Svg/consts';
@@ -118,8 +120,16 @@ class NavigateAppPage extends BasePage {
     await (await this.radioGroupPage).click();
   }
 
+  async clickAndGoToRadioGroupExperimentalPage() {
+    await (await this.radioGroupExperimentalPage).click();
+  }
+
   async clickAndGoToSeparatorPage() {
     await (await this.separatorPage).click();
+  }
+
+  async clickAndGoToShadowPage() {
+    await (await this.shadowPage).click();
   }
 
   async clickAndGoToShimmerPage() {
@@ -254,8 +264,16 @@ class NavigateAppPage extends BasePage {
     return By(HOMEPAGE_RADIOGROUP_BUTTON);
   }
 
+  private get radioGroupExperimentalPage() {
+    return By(HOMEPAGE_RADIO_GROUP_EXPERIMENTAL_BUTTON);
+  }
+
   private get separatorPage() {
     return By(HOMEPAGE_SEPARATOR_BUTTON);
+  }
+
+  private get shadowPage() {
+    return By(HOMEPAGE_SHADOW_BUTTON);
   }
 
   private get shimmerPage() {