+++ apps/fluent-tester/src/TestComponents/Badge/BasicBadgeTest.tsx
@@ -15,12 +15,11 @@ import {
   BadgeProps,
   BadgeTokens,
 } from '@fluentui-react-native/badge';
-import { SvgIconProps } from '@fluentui-react-native/icon';
 import { StyledPicker } from '../Common/StyledPicker';
 import { satyaPhotoUrl } from './../PersonaCoin/styles';
-import TestSvg from '../../FluentTester/test-data/test.svg';
 import { ToggleButton } from '@fluentui/react-native';
 import { useFluentTheme } from '@fluentui-react-native/framework';
+import { svgProps, iconProps } from '../Common/iconExamples';
 
 const badgeColors: BadgeColor[] = [...BadgeColors];
 const badgeShapes: BadgeShape[] = [...BadgeShapes];
@@ -50,10 +49,6 @@ export const BasicBadge: React.FunctionComponent = () => {
   const onShowIconChange = useCallback(() => setShowIcon(!showIcon), [showIcon, setShowIcon]);
 
   const theme = useFluentTheme();
-  const svgProps: SvgIconProps = {
-    src: TestSvg,
-    viewBox: '0 0 500 500',
-  };
 
   const fontBuiltInProps = {
     fontFamily: 'Arial',
@@ -62,7 +57,6 @@ export const BasicBadge: React.FunctionComponent = () => {
   };
 
   const svgIconsEnabled = ['ios', 'macos', 'win32', 'android'].includes(Platform.OS as string);
-  const iconProps = { svgSource: svgProps, width: 20, height: 20 };
 
   const shadow = showShadow ? theme.shadows.shadow4 : undefined;
   const CustomBadge = useCustomizedBadge({ shadowToken: shadow });
@@ -126,12 +120,12 @@ export const BasicBadge: React.FunctionComponent = () => {
       {svgIconsEnabled && (
         <>
           <Text>Badge with icon</Text>
-          <Badge icon={{ svgSource: svgProps }} iconPosition="after">
+          <Badge icon={iconProps} iconPosition="after">
             Badge with
             <Image source={{ uri: satyaPhotoUrl }} style={{ width: 20, height: 20 }} />
             <Text style={{ backgroundColor: 'yellow' }}>optional content</Text>
           </Badge>
-          <Badge appearance="outline" icon={iconProps} />
+          <Badge appearance="outline" icon={{ ...iconProps, width: 20, height: 20 }} />
           <Badge icon={{ fontSource: { ...fontBuiltInProps }, color: '#fff' }}>Badge with icon</Badge>
           <Text>Customized Badge with icon</Text>
           <StyledBadge icon={{ svgSource: svgProps }}>Styled badge</StyledBadge>