+++ apps/fluent-tester/src/TestComponents/Button/ButtonSizeTestSection.tsx
@@ -1,16 +1,12 @@
-import { ButtonV1 as Button, CompoundButton } from '@fluentui/react-native';
-import { SvgIconProps } from '@fluentui-react-native/icon';
+import { ButtonV1 as Button, CompoundButton, FAB } from '@fluentui/react-native';
 import * as React from 'react';
 import { Platform, View } from 'react-native';
 import { commonTestStyles, testContentRootViewStyle } from '../Common/styles';
-import TestSvg from './test.svg';
+import { iconProps } from '../Common/iconExamples';
 
 export const ButtonSizeTest: React.FunctionComponent = () => {
-  const svgProps: SvgIconProps = {
-    src: TestSvg,
-    viewBox: '0 0 500 500',
-  };
   const svgIconsEnabled = ['ios', 'macos', 'win32', 'android'].includes(Platform.OS as string);
+
   return (
     <View style={testContentRootViewStyle}>
       {svgIconsEnabled && (
@@ -18,7 +14,7 @@ export const ButtonSizeTest: React.FunctionComponent = () => {
           <Button
             iconOnly
             size="small"
-            icon={{ svgSource: svgProps }}
+            icon={iconProps}
             accessibilityLabel="Small size button with accessibility icon"
             style={commonTestStyles.vmargin}
             tooltip="button tooltip"
@@ -26,26 +22,32 @@ export const ButtonSizeTest: React.FunctionComponent = () => {
           <Button
             iconOnly
             size="medium"
-            icon={{ svgSource: svgProps }}
+            icon={iconProps}
             accessibilityLabel="Medium size button with accessibility icon"
             style={commonTestStyles.vmargin}
           />
           <Button
             iconOnly
             size="large"
-            icon={{ svgSource: svgProps }}
+            icon={iconProps}
             accessibilityLabel="Large size button with accessibility icon"
             style={commonTestStyles.vmargin}
           />
-          <Button size="small" icon={{ svgSource: svgProps }} style={commonTestStyles.vmargin}>
+          <Button size="small" icon={iconProps} style={commonTestStyles.vmargin}>
             Small Button with icon
           </Button>
-          <Button size="medium" icon={{ svgSource: svgProps }} style={commonTestStyles.vmargin}>
+          <Button size="medium" icon={iconProps} style={commonTestStyles.vmargin}>
             Medium Button with icon
           </Button>
-          <Button size="large" icon={{ svgSource: svgProps }} style={commonTestStyles.vmargin}>
+          <Button size="large" icon={iconProps} style={commonTestStyles.vmargin}>
             Large Button with icon
           </Button>
+          <FAB size="small" icon={iconProps} style={commonTestStyles.vmargin}>
+            Small FAB
+          </FAB>
+          <FAB size="large" icon={iconProps} style={commonTestStyles.vmargin}>
+            Large FAB
+          </FAB>
         </>
       )}
       <Button size="small" style={commonTestStyles.vmargin}>
@@ -57,35 +59,39 @@ export const ButtonSizeTest: React.FunctionComponent = () => {
       <Button size="large" style={commonTestStyles.vmargin}>
         Large
       </Button>
-      <Button loading size="small" style={commonTestStyles.vmargin}>
-        Loading Button Small
-      </Button>
-      <Button loading size="medium" style={commonTestStyles.vmargin}>
-        Loading Button Medium
-      </Button>
-      <Button loading size="large" style={commonTestStyles.vmargin}>
-        Loading Button Large
-      </Button>
-      <CompoundButton secondaryContent="Small compound button" size="small" style={commonTestStyles.vmargin}>
-        Compound Button
-      </CompoundButton>
-      <CompoundButton secondaryContent="Medium compound button" size="medium" style={commonTestStyles.vmargin}>
-        Compound Button
-      </CompoundButton>
-      <CompoundButton secondaryContent="Large compound button" size="large" style={commonTestStyles.vmargin}>
-        Compound Button
-      </CompoundButton>
-      {svgIconsEnabled && (
+      {Platform.OS !== 'android' && (
         <>
-          <CompoundButton icon={{ svgSource: svgProps }} secondaryContent="SecondaryContent" size="small" style={commonTestStyles.vmargin}>
-            Content
+          <Button loading size="small" style={commonTestStyles.vmargin}>
+            Loading Button Small
+          </Button>
+          <Button loading size="medium" style={commonTestStyles.vmargin}>
+            Loading Button Medium
+          </Button>
+          <Button loading size="large" style={commonTestStyles.vmargin}>
+            Loading Button Large
+          </Button>
+          <CompoundButton secondaryContent="Small compound button" size="small" style={commonTestStyles.vmargin}>
+            Compound Button
           </CompoundButton>
-          <CompoundButton icon={{ svgSource: svgProps }} secondaryContent="SecondaryContent" size="medium" style={commonTestStyles.vmargin}>
-            Content
+          <CompoundButton secondaryContent="Medium compound button" size="medium" style={commonTestStyles.vmargin}>
+            Compound Button
           </CompoundButton>
-          <CompoundButton icon={{ svgSource: svgProps }} secondaryContent="SecondaryContent" size="large" style={commonTestStyles.vmargin}>
-            Content
+          <CompoundButton secondaryContent="Large compound button" size="large" style={commonTestStyles.vmargin}>
+            Compound Button
           </CompoundButton>
+          {svgIconsEnabled && (
+            <>
+              <CompoundButton icon={iconProps} secondaryContent="SecondaryContent" size="small" style={commonTestStyles.vmargin}>
+                Content
+              </CompoundButton>
+              <CompoundButton icon={iconProps} secondaryContent="SecondaryContent" size="medium" style={commonTestStyles.vmargin}>
+                Content
+              </CompoundButton>
+              <CompoundButton icon={iconProps} secondaryContent="SecondaryContent" size="large" style={commonTestStyles.vmargin}>
+                Content
+              </CompoundButton>
+            </>
+          )}
         </>
       )}
     </View>