+++ apps/fluent-tester/src/TestComponents/MenuButtonExperimental/StandardMenuButtonTest.tsx
@@ -1,10 +1,11 @@
 import * as React from 'react';
 import { Separator, ContextualMenuProps } from '@fluentui/react-native';
 import { Text, View, Switch, Platform } from 'react-native';
-import { menuItems, iconProps, testImage } from './testData';
+import { menuItems, iconProps } from './testData';
 import { viewWrapperStyle, columnStyle, rowStyle, textColor } from './MenuButtonTestStyles';
 import { IconSourcesType } from '@fluentui-react-native/icon';
 import { MenuButton } from '@fluentui-react-native/experimental-menu-button';
+import { testImage } from '../Common/iconExamples';
 
 export const StandardMenuButton: React.FunctionComponent = () => {
   const [lastMenuItemClicked, setLastMenuItemClicked] = React.useState(null);