+++ apps/fluent-tester/src/TestComponents/RadioGroupExperimental/consts.ts
@@ -1,2 +1,33 @@
 export const HOMEPAGE_RADIO_GROUP_EXPERIMENTAL_BUTTON = 'Homepage_RadioGroupExperimental_Button';
 export const RADIO_GROUP_EXPERIMENTAL_TESTPAGE = 'RadioGroupExperimental_TestPage';
+
+/********************************************************************/
+/********************* E2E Testing RadioGroup 1 *********************/
+/********************************************************************/
+export const RADIOGROUP_TEST_COMPONENT = 'RadioGroupExperimental_Test_Component';
+export const RADIOGROUP_ACCESSIBILITY_LABEL = 'E2E testing RadioGroup accessibility label';
+
+/* E2E Testing RadioButton 1 (1 of 4) */
+export const FIRST_RADIO = 'First_Radio';
+export const FIRST_RADIO_ACCESSIBILITY_LABEL = 'E2E testing Radio accessibility label';
+
+/* E2E Testing RadioButton 1 (2 of 4) No accessibilityLabel */
+export const SECOND_RADIO = 'Second_Radio';
+export const SECOND_RADIO_LABEL = 'Test RadioGroup2 - No Accessibility Label';
+
+/* E2E Testing RadioButton 1 (3 of 4) */
+export const THIRD_RADIO = 'Third_Radio';
+
+/* E2E Testing RadioButton 1 (4 of 4) */
+export const FOURTH_RADIO = 'Fourth_Radio';
+
+/* E2E Testing RadioButton 2 (5 of 8) */
+export const FIFTH_RADIO = 'Fifth_Radio';
+
+/********************************************************************/
+/********************* E2E Testing RadioGroup 2 *********************/
+/********************************************************************/
+export const RADIOGROUP_NO_A11Y_LABEL_COMPONENT = 'RadioGroup_No_A11y_label_Component';
+export const RADIOGROUP_TEST_COMPONENT_LABEL = 'Test RadioGroupExperimental2 - No Accessibility Label'; // A component on each specific test page
+
+export const RADIOGROUP_ON_PRESS = 'RadioGroup_On_Press'; // For testing the press functionality on a button